requirements.txt
azure-storage-blob
rasterio